refactor: consolidate polygon boolean algorithm files
